+++ nym-vpn-app/src/assets/onboarding/index.ts
@@ -1,0 +1,6 @@
+/// <reference types="vite-plugin-svgr/client" />
+
+export { default as Speed } from './speed.svg?react';
+export { default as StopTracking } from './stop-tracking.svg?react';
+export { default as ZeroKnowledge } from './zero-knowledge.svg?react';
+export { default as Welcome } from './welcome.svg?react';