+++ nym-vpn-app/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add Privy to allow using social accounts for log in
+- Add new onboarding screens
 
 ## [1.21.0] - 2025-12-18